Rename "predicates" directory to "search"
Resolving one conversation in the review of #1237.
