+++ chatterino.pro
@@ -122,9 +122,9 @@ SOURCES += \
     src/messages/MessageColor.cpp \
     src/messages/MessageContainer.cpp \
     src/messages/MessageElement.cpp \
-    src/messages/predicates/AuthorPredicate.cpp \
-    src/messages/predicates/LinkPredicate.cpp \
-    src/messages/predicates/SubstringPredicate.cpp \
+    src/messages/search/AuthorPredicate.cpp \
+    src/messages/search/LinkPredicate.cpp \
+    src/messages/search/SubstringPredicate.cpp \
     src/providers/bttv/BttvEmotes.cpp \
     src/providers/bttv/LoadBttvChannelEmote.cpp \
     src/providers/chatterino/ChatterinoBadges.cpp \
@@ -288,10 +288,10 @@ HEADERS += \
     src/messages/MessageContainer.hpp \
     src/messages/MessageElement.hpp \
     src/messages/MessageParseArgs.hpp \
-    src/messages/predicates/AuthorPredicate.hpp \
-    src/messages/predicates/LinkPredicate.hpp \
-    src/messages/predicates/MessagePredicate.hpp \
-    src/messages/predicates/SubstringPredicate.hpp \
+    src/messages/search/AuthorPredicate.hpp \
+    src/messages/search/LinkPredicate.hpp \
+    src/messages/search/MessagePredicate.hpp \
+    src/messages/search/SubstringPredicate.hpp \
     src/messages/Selection.hpp \
     src/PrecompiledHeader.hpp \
     src/providers/bttv/BttvEmotes.hpp \
